100% fixed
 Era um erro de lógica de calculo. No serviço ContractService lá no calculo do PaymentFee, eu estava apenas passando o basicQuota sem somar com o Interest.

agora sim, a aplicação está 100% funcional.

+++ ex16_1/services/PaypalService.java
@@ -9,9 +9,6 @@ public class PaypalService implements OnlinePaymentService {
     public double paymentFee(double amount) {
         return amount * FEE_PERCENTAGE;
     }
-
     @Override
-    public double interest(double amount, int months) {
-        return amount * MONTHLY_INTEREST * months;
-    }
+    public double interest(double amount, int months) { return amount * MONTHLY_INTEREST * (double)months; }
 }
